* Fix, thanks Piotrek1447
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@3102 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ creature.cpp
@@ -706,7 +706,7 @@ bool Creature::onDeath()
 		if(!it->getKillerCreature()->onKilledCreature(this, flags) && lastHit)
 			return false;
 
-		int8_t value = 0;
+		int16_t value = 0;
 		if(hasBitSet((uint32_t)KILLFLAG_UNJUSTIFIED, flags))
 			value += 1;
